Move most of MusicPlayer's functionality into PlayQueue
This is in order to fix the bug where we aren't able to differentiate from the first render of a song and the second time it is clicked.
